requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
Deprecated==1.2.14
Django==4.2.6
django-encrypted-id-cryptography==1.1.0
django-environ==0.11.2
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
jwcrypto==1.5.0
oauthlib==3.2.2
packaging==24.0
Pillow==10.1.0
psycopg2==2.9.9
pycparser==2.21
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
whitenoise==6.6.0
wrapt==1.15.0
